backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
sqlalchemy==2.0.27
aiosqlite==0.19.0
python-jose==3.3.0
passlib==1.7.4
python-dotenv==1.0.1